BFS ALGORITHM
Breadth-First Search(also known as Breadth-First Traversal) is a traversal mechanism that, as the name might suggest, traverses the graph in a breadthwise manner.

One of them is finding the shortest path/distance between two graph nodes. We start our BFS from one of the two nodes and explore its neighboring nodes.